P2 mcp: preserve WebSocket batch array width
Finding: MEM-web-mcp-adapters-002.

Fault: the WebSocket JSON-RPC batch path explicitly narrowed json_object_array_length() from size_t to int and drove json_object_array_get_idx() with an int loop index. HTTP and SSE already preserve the json-c native width.

Reachability: json-c 0.14 through the bundled 0.17 and current releases retain an INT32_MAX input limit in json_tokener_parse_verbose(). Minimum valid array syntax bounds the current parsed count to 1,073,741,823, below INT_MAX, so truncation is potential rather than currently triggerable. Uncompressed WebSocket fragmentation has no aggregate cap, but requests above the parser limit fail before this loop.

Caller and framing contracts: websocket-handshake installs the callback and websocket_client_process_message invokes it synchronously with a complete NUL-terminated, UTF-8-validated text message, then resets message storage after return. The callback does not retain the message. Frame, compression, parser, and batch limits are unchanged.

Solution: use size_t for both the json-c array length and loop index, matching the dependency API and sibling MCP transports.

Functional impact: none for current valid input. Parse errors, empty and notification-only batches, mixed-batch response filtering, response ordering, total-length construction, cleanup, disconnect behavior, and sent JSON remain unchanged.

Performance impact: none. The same O(n) loop performs the same lookups, dispatches, allocations, and sends with no added branches, scans, syscalls, or locks.

Concurrency and ownership impact: none. Only stack-local scalar widths change; request ownership, response-buffer ownership, synchronous send copying, close/disconnect context release, and owner-thread execution are unchanged.

Validation: focused low-priority translation-unit compile passed; git diff --check passed; supported json-c API/parser sources and ABI bounds were verified; the canonical blocking gate passed with structured approvals from glm, minimax, deepseek, kimi, qwen, and mimo.
(cherry picked from commit aedfabb35f3f715f0e94f70b726aa77442e68082)

+++ src/web/mcp/adapters/mcp-websocket.c
@@ -111,12 +111,12 @@ void mcp_websocket_on_message(struct websocket_server_client *wsc, const char *m
     }
 
     if (json_object_is_type(request, json_type_array)) {
-        int len = (int)json_object_array_length(request);
+        size_t len = json_object_array_length(request);
         BUFFER **responses = NULL;
         size_t responses_used = 0;
         size_t responses_size = 0;
 
-        for (int i = 0; i < len; i++) {
+        for (size_t i = 0; i < len; i++) {
             struct json_object *req_item = json_object_array_get_idx(request, i);
             BUFFER *resp_item = mcp_jsonrpc_process_single_request(mcpc, req_item, NULL);
             if (resp_item) {
